pyproject.toml
[project]
name = "multidocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "llama-index-core>=0.14.19",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-llms-openai>=0.7.4",
    "llama-index-tools-mcp>=0.4.8",
    "llama-index-vector-stores-postgres>=0.8.1",
    "pgvector>=0.4.2",
    "psycopg>=3.3.3",
    "psycopg-binary>=3.3.3",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
    "python-telegram-bot>=22.7",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "wsproto>=1.3.2",
]
